+++ yosai/core/mgt/mgt_settings.py
@@ -71,6 +71,12 @@ class SecurityManagerSettings:
                     authc_verifiers_cls = tuple([maybe_resolve(authc_verifiers)(self.settings)])
                 verifiers['authc_verifiers'] = authc_verifiers_cls
 
+            authz_verifiers = realm_attributes.get('authz_verifiers')
+            if authz_verifiers:
+                permission_verifier_cls = authz_verifiers.get('permission_verifier')
+                if permission_verifier_cls:
+                    verifiers['permission_verifier'] = maybe_resolve(permission_verifier_cls)()
+
             realms.append([realm_cls, account_store_cls, verifiers])
 
         return realms